[Feature/FP-24] 도메인 관련 테스트 코드 추가

+++ finpik-common/src/main/java/finpik/dto/CreateRecommendedLoanProductEvent.java
@@ -15,11 +15,11 @@ public record CreateRecommendedLoanProductEvent(
         String productName,
         Float minInterestRate,
         Float maxInterestRate,
-        Integer loanLimitAmount
+        Long loanLimitAmount
     ) {
         public static CreateRecommendedLoanProductEventContent of(
             Long loanProductId, String productName, Float minInterestRate,
-            Float maxInterestRate, Integer loanLimitAmount
+            Float maxInterestRate, Long loanLimitAmount
         ) {
             return new CreateRecommendedLoanProductEventContent(
                 loanProductId,
